client: Allow overriding global service handlers
The library was built to make it easy to plug handlers in for
services, but this sadly assumed a very global nature for such
handlers. Which is fine for many applications, but sometimes you
want two different monitors to handle different sets of services,
so it should also be possible to tell the Monitor objects more
directly what services it may detect and monitor.