service: onchange can be an awaitable
When onchange is an awaitable, a dbus SetValue call is scheduled
as a task that runs later, and can return much later. The task
is then responsible for updating the internal value and sending
ItemsChanged.

This allows building services that only update their value once
the value was accepted by the hardware.