client: hook to allow overriding Item inner class
This allows extending the Item-tracking class used internally
and where the Service-tracking class is extended, using the
extended subclass. This in turn allows adding more functionality
in projects without adding all the heavy stuff into the
base library.

Specifically: This is so the Item can be extended to wait for
a value to become valid, a fairly common use case.